requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
nats-py
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
python-dotenv==1.0.0
